// src/domains/tax-receipts/dto/update-tax-receipt.dto.ts
import { Types } from 'mongoose';

import { TaxReceiptType } from '../models';
import type { TaxReceiptMetadata } from '../models';

export class UpdateTaxReceiptDto {
  donationId?: Types.ObjectId;

  donorUserId?: Types.ObjectId;

  institutionId?: Types.ObjectId;

  receiptNumber?: string;

  type?: TaxReceiptType;

  amount?: number;

  issuedAt?: Date;

  documentUrl?: string;

  metadata?: TaxReceiptMetadata;
}